file-upload

+++ controllers/admin.js
@@ -1,3 +1,5 @@
+const fileHelper = require('../util/file');
+
 const { validationResult } = require('express-validator');
 const Product = require('../models/product');
 
@@ -14,10 +16,27 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title;
-  const imageUrl = req.body.imageUrl;
+  const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
+
   const errors = validationResult(req);
+
+  if (!image) {
+    return res.status(422).render('admin/edit-product', {
+      pageTitle: 'Add Product',
+      path: '/admin/add-product',
+      editing: false,
+      hasError: true,
+      product: {
+        title: title,
+        price: price,
+        description: description,
+      },
+      errorMessage: 'Attached file is not an image',
+      validationErrors: [],
+    });
+  }
 
   if (!errors.isEmpty()) {
     return res.status(422).render('admin/edit-product', {
@@ -29,12 +48,14 @@ exports.postAddProduct = (req, res, next) => {
         title: title,
         price: price,
         description: description,
-        imageUrl: imageUrl,
       },
       errorMessage: errors.array()[0].msg,
       validationErrors: errors.array(),
     });
   }
+
+  const imageUrl = image.path;
+
   const product = new Product({
     title: title,
     price: price,
@@ -50,9 +71,90 @@ exports.postAddProduct = (req, res, next) => {
       res.redirect('/admin/products');
     })
     .catch((err) => {
-      console.log(err);
+      // res.redirect('/500');
+      const error = new Error(err);
+      error.httpStatusCode = 500;
+      return next(error);
     });
 };
+
+// exports.postAddProduct = (req, res, next) => {
+//   const title = req.body.title;
+//   const image = req.file;
+//   const price = req.body.price;
+//   const description = req.body.description;
+//   const errors = validationResult(req);
+
+//   if (!image) {
+//     return res.status(422).render('admin/edit-product', {
+//       pageTitle: 'Add Product',
+//       path: '/admin/add-product',
+//       editing: false,
+//       hasError: true,
+//       product: {
+//         title: title,
+//         price: price,
+//         description: description,
+//       },
+//       errorMessage: 'Attached file is not an image',
+//       validationErrors: [],
+//     });
+//   }
+
+//   if (!errors.isEmpty()) {
+//     console.log(errors.array());
+//     return res.status(422).render('admin/edit-product', {
+//       pageTitle: 'Add Product',
+//       path: '/admin/add-product',
+//       editing: false,
+//       hasError: true,
+//       product: {
+//         title: title,
+//         imageUrl: imageUrl,
+//         price: price,
+//         description: description,
+//       },
+//       errorMessage: errors.array()[0].msg,
+//       validationErrors: errors.array(),
+//     });
+//   }
+
+//   const product = new Product({
+//     // _id: new mongoose.Types.ObjectId('5badf72403fd8b5be0366e81'),
+//     title: title,
+//     price: price,
+//     description: description,
+//     imageUrl: imageUrl,
+//     userId: req.user,
+//   });
+//   product
+//     .save()
+//     .then((result) => {
+//       // console.log(result);
+//       console.log('Created Product');
+//       res.redirect('/admin/products');
+//     })
+//     .catch((err) => {
+//       // return res.status(500).render('admin/edit-product', {
+//       //   pageTitle: 'Add Product',
+//       //   path: '/admin/add-product',
+//       //   editing: false,
+//       //   hasError: true,
+//       //   product: {
+//       //     title: title,
+//       //     imageUrl: imageUrl,
+//       //     price: price,
+//       //     description: description
+//       //   },
+//       //   errorMessage: 'Database operation failed, please try again.',
+//       //   validationErrors: []
+//       // });
+//       // res.redirect('/500');
+//       const error = new Error(err);
+//       error.httpStatusCode = 500;
+//       return next(error);
+//     });
+// };
 
 exports.getEditProduct = (req, res, next) => {
   const editMode = req.query.edit;
@@ -75,14 +177,18 @@ exports.getEditProduct = (req, res, next) => {
         validationErrors: [],
       });
     })
-    .catch((err) => console.log(err));
+    .catch((err) => {
+      const error = new Error(err);
+      error.httpStatusCode = 500;
+      return next(error);
+    });
 };
 
 exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId;
   const updatedTitle = req.body.title;
   const updatedPrice = req.body.price;
-  const updatedImageUrl = req.body.imageUrl;
+  const image = req.file;
   const updatedDesc = req.body.description;
   const errors = validationResult(req);
 
@@ -96,7 +202,7 @@ exports.postEditProduct = (req, res, next) => {
         title: updatedTitle,
         price: updatedPrice,
         description: updatedDesc,
-        imageUrl: updatedImageUrl,
+
         _id: prodId,
       },
       errorMessage: errors.array()[0].msg,
@@ -112,34 +218,55 @@ exports.postEditProduct = (req, res, next) => {
       product.title = updatedTitle;
       product.price = updatedPrice;
       product.description = updatedDesc;
-      product.imageUrl = updatedImageUrl;
+      if (image) {
+        fileHelper.deleteFile(product.imageUrl);
+        product.imageUrl = image.path;
+      }
       return product.save().then((result) => {
         console.log('UPDATED PRODUCT!');
         res.redirect('/admin/products');
       });
     })
-    .catch((err) => console.log(err));
+    .catch((err) => {
+      const error = new Error(err);
+      error.httpStatusCode = 500;
+      return next(error);
+    });
 };
 
 exports.getProducts = (req, res, next) => {
   Product.find({ userId: req.user._id })
     .then((products) => {
-      // console.log(products);
       res.render('admin/products', {
         prods: products,
         pageTitle: 'Admin Products',
         path: '/admin/products',
       });
     })
-    .catch((err) => console.log(err));
+    .catch((err) => {
+      const error = new Error(err);
+      error.httpStatusCode = 500;
+      return next(error);
+    });
 };
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  Product.findById(prodId)
+    .then((product) => {
+      if (!product) {
+        return next(new Error('Product not found'));
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id });
+    })
     .then(() => {
       console.log('DESTROYED PRODUCT');
       res.redirect('/admin/products');
     })
-    .catch((err) => console.log(err));
+    .catch((err) => {
+      const error = new Error(err);
+      error.httpStatusCode = 500;
+      return next(error);
+    });
 };
